requirements.txt
streamlit==1.48.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.52
langchain-openai==0.3.13
langgraph==0.3.31
langsmith==0.3.45

chromadb==1.0.20
cohere==4.51
openai==1.75.0
tiktoken==0.9.0

python-dotenv==1.1.0
pydantic==2.11.3
requests==2.32.3
tenacity==9.1.2
orjson==3.10.16

watchfiles==1.1.0
